1. restructure the project structure
