+++ src/core/engine/mod.rs
@@ -1,3 +1,2 @@
 pub mod engine;
-mod watcher;
-pub mod init;
+pub mod initialization;
