+++ src/core/environment/interface.rs
@@ -1,4 +1,4 @@
-use crate::core::env::java::check_java_version;
+use crate::core::environment::java::check_java_version;
 use crate::entity::flow::Env;
 
 pub fn check(name: Vec<Env>) {
