+++ src/core/environment/java.rs
