+++ src/core/extension/interface.rs
@@ -1,8 +1,9 @@
 use serde_json::from_str;
+
 use crate::conf::runtime::get_runtime_conf;
-use crate::core::common::log::interface::warn;
 use crate::core::extension::dll::interface::call_dll_extension_method;
 use crate::entity::ext::Extension;
+use crate::tools::log::interface::warn;
 
 pub fn call(extension_name: String, function_name: String) {
     let extension: Extension = from_str(get_runtime_conf(format!("ext_{}", extension_name).as_str()).unwrap().as_str()).unwrap();
