+++ src/core/flow/controller/interface.rs
@@ -1,9 +1,9 @@
 use std::path::Path;
 
-use crate::core::common::log::interface::info;
 use crate::core::flow::controller::check::check_require;
 use crate::core::flow::exec::flow::exec_standardisation_flow;
 use crate::core::flow::resolver::interface::flow_resolver;
+use crate::tools::log::interface::info;
 
 // 这个东西其实就是Json（目前来说），后续可能会进行一些加密的操作
 pub async fn exec_fl_flow(path: &Path) {
