+++ src/core/flow/exec/node.rs
@@ -1,7 +1,7 @@
-use crate::core::common::log::interface::debug;
 use crate::core::flow::expression::interface::resolve_said_expression;
 use crate::core::flow::handler::interface::handler;
 use crate::entity::flow::{FlowData, Node};
+use crate::tools::log::interface::debug;
 
 // Node 调度
 // Node 需要对应的Handler执行
