+++ src/core/flow/expression/said/resolve_num.rs
@@ -1,5 +1,5 @@
-use crate::core::common::log::interface::debug;
 use crate::core::flow::expression::said::entity::Expression;
+use crate::tools::log::interface::debug;
 
 pub fn resolve_num(expression: &Expression) {
     debug(format!("{:?}", expression).as_str())
