+++ src/core/flow/handler/net/interface.rs
@@ -1,6 +1,6 @@
-use crate::core::common::log::interface::warn;
 use crate::core::flow::handler::net::http::handle_net_http;
 use crate::entity::flow::{FlowData, Node};
+use crate::tools::log::interface::warn;
 
 pub async fn handle_net(node: Node, flow_data: &mut FlowData) {
     let handler_path: Vec<_> = node.handler.split(".").collect();
