+++ src/core/flow/resolver/interface.rs
@@ -1,8 +1,8 @@
 use std::path::Path;
 
-use crate::core::common::log::interface::warn;
 use crate::core::flow::resolver::flow::resolver_flow;
 use crate::entity::flow::Flow;
+use crate::tools::log::interface::warn;
 
 pub fn flow_resolver(path: &Path) -> Flow {
     if let Some(extension) = path.extension() {
