+++ src/core/mod.rs
@@ -1,7 +1,6 @@
 pub mod engine;
 pub mod flow;
-mod env;
-pub mod common;
+mod environment;
 mod event;
 pub mod script;
 pub mod extension;
