+++ src/core/runtime/engine.rs
@@ -12,12 +12,6 @@ pub fn set_engine_info(key: &str, value: &str) {
     data.insert(key.to_string(), value.to_string());
 }
 
-// 获取指定运行时配置
-pub fn get_runtime_conf(key: &str) -> Option<String> {
-    let data = RUNTIME_ENGINE.lock().unwrap();
-    data.get(key).cloned()
-}
-
 // 修改指定运行时配置
 pub fn set_runtime_conf(key: &str, value: &str) {
     let mut data = RUNTIME_ENGINE.lock().unwrap();
