+++ src/core/runtime/extension.rs
@@ -1,0 +1,1 @@
+// 此runtime主要由插件使用
