+++ src/core/runtime/mod.rs
@@ -2,3 +2,4 @@ pub mod engine;
 pub mod flow;
 pub mod script;
 pub mod extension;
+pub mod config;
