+++ src/core/script/bat.rs
@@ -2,7 +2,7 @@ use std::env;
 use std::path::Path;
 use std::process::Command;
 
-use crate::core::common::log::interface::{info, script_fail, script_log};
+use crate::tools::log::interface::{info, script_fail, script_log};
 
 pub fn exec_bat_script(path: &Path) {
     info(format!("Find bat in path -> {:?}", path).as_str());
