+++ src/entity/mod.rs
@@ -2,4 +2,4 @@ pub mod flow;
 pub mod net;
 pub mod simx;
 pub mod ext;
-pub mod config;
+pub mod engine;
