+++ src/tools/log/direct/shell/basic.rs
