+++ src/tools/log/direct/shell/colourful.rs
