+++ src/tools/mod.rs
@@ -1,2 +1,3 @@
 pub mod http;
 pub mod bytes;
+pub mod log;
